[#56]refactor: Update firmwareSimulator to store the dodecahedron as JSON
The dodecahedron is now represented as a JSON object in the dodecahedron/dodecahedron.json file. This was done to allow the user to see changes made to the device via updates sent by the user via AWS. The user can now see what side has a given habit assoicated with it.