[#64]chore:Remove device timestamp from data being sent by the simulator when a TIME habit is executed.
Removed timestamp from being sent when TIME habit is sent. However, the corresponding protofiles have not been updated. Added important links and sources to the setup.py file. Added links for the certificate files of firmwareSimulator to the config.py file as comments. These are the ones currently used, while the certificate2 filepaths are used during development of the setup.py file.