[#79]fix: Change file mode from 'a' to 'w'
This was done to reduce the steps necessary to undo the setup. Instead of going into each of the files in the certificates folder
and deleting their contents (one cannot simply delete the files because they contain secret keys that should not be forgotten in the PCs recycle bin).
Now when running the setup.py again the contents of the files get simply overwritten.